fix: structural MEDIUM cap in evaluator, shared guidancePriority
Two review findings addressed:

1. Device posture MEDIUM cap now enforced in buildFinding() as a
   single structural chokepoint. Remote rules with level: critical
   for device_posture are clamped to medium at evaluation time,
   making every downstream consumer automatically safe.

2. guidancePriority() extracted to shared GuidanceUtils object,
   eliminating divergent behavior between ReportFormatter (missing
   REVIEW prefix) and TimelineExporter. Both now delegate to the
   same function with unified priority levels.
